fix export users to do it only on migration

+++ docker-stacks/02-keycloak/etc/hooks/before-stop.d/10-ExportUsers.sh
@@ -1,16 +1,25 @@
-# Check if the container is running
-keycloakContainer=$(echo $(docker ps --format '{{.Names}}' | grep "keycloak-1"))
-if [ ! $keycloakContainer = "" ]; then
-    echo "Keycloak container is running. Launching export of users..."
-    if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 18 ]]; then 
-        rm -rf ${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/*
-        docker compose exec keycloak /opt/keycloak/bin/kc.sh export --dir "/opt/keycloak/data/export/" --users different_files --users-per-file 100 --realm ${SIMVA_SSO_REALM:-simva} --optimized
+if [[ -e "${SIMVA_PROJECT_DIR}/.simva-initialized" ]]; then 
+    echo "SIMVA is initialized." 
+    migrationinProgressFile="${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/.migrationinprogress"
+    if [[ -e "$migrationinProgressFile" ]]; then
+            echo "Migration in progress. Exporting realm..."
+            # Check if the container is running
+            keycloakContainer=$(echo $(docker ps --format '{{.Names}}' | grep "keycloak-1"))
+            if [ ! $keycloakContainer = "" ]; then
+                echo "Keycloak container is running. Launching export of users..."
+                if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 18 ]]; then 
+                    rm -rf ${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/*
+                    docker compose exec keycloak /opt/keycloak/bin/kc.sh export --dir "/opt/keycloak/data/export/" --users different_files --users-per-file 100 --realm ${SIMVA_SSO_REALM:-simva} --optimized
+                else 
+                    #if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 13 ]]; then
+                    #    docker exec $container_name /opt/jboss/tools/docker-entrypoint.sh -Dkeycloak.migration.action=export -Dkeycloak.migration.provider=dir -Dkeycloak.migration.dir=/var/tmp/simva-realm -Dkeycloak.migration.usersExportStrategy=SAME_FILE
+                    #fi
+                    echo "Please upgrade to a newer keycloak version ( > 18.*.*) before exporting users and realm."
+                fi;   
+            else
+                echo "Keycloak container is not running."
+            fi
     else 
-        #if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 13 ]]; then
-        #    docker exec $container_name /opt/jboss/tools/docker-entrypoint.sh -Dkeycloak.migration.action=export -Dkeycloak.migration.provider=dir -Dkeycloak.migration.dir=/var/tmp/simva-realm -Dkeycloak.migration.usersExportStrategy=SAME_FILE
-        #fi
-        echo "Please upgrade to a newer keycloak version ( > 18.*.*) before exporting users and realm."
-    fi;   
-else
-    echo "Keycloak container is not running."
+        echo "Nothing to do. No migration detected."
+    fi
 fi
